Moved logic from Querier to MsgVariants
